Generalize messaging class for use with SC integration.
