+++ src/outlet/index.ts
@@ -1,17 +1,14 @@
 import { rawTokenCheck, readMagicUrl, storeMagicURL } from "../core";
-import { logger, requiredParams } from "../utils/index";
-import { decodeTokens, filterTokens } from "./../core/index";
-import {
-	MessageAction,
-	MessageResponseInterface,
-	MessageResponseAction,
-} from "../client/messaging";
+import { logger, requiredParams } from "../utils";
+import { decodeTokens, filterTokens } from "../core";
+import { OutletAction, OutletResponseAction } from "../client/messaging";
 import { AuthHandler } from "./auth-handler";
 
 // requred for default TicketDecoder
 import { SignedDevconTicket } from "@tokenscript/attestation/dist/asn1/shemas/SignedDevconTicket";
 import { AsnParser } from "@peculiar/asn1-schema";
-import { uint8toBuffer } from "./../utils/index";
+import { uint8toBuffer } from "../utils";
+import { ResponseActionBase, ResponseInterfaceBase } from "../core/messaging";
 
 interface OutletInterface {
 	config: any;
@@ -75,12 +72,12 @@ export class Outlet {
 		// TODO: should issuer be validated against requested issuer?
 
 		switch (action) {
-		case MessageAction.GET_ISSUER_TOKENS: {
+		case OutletAction.GET_ISSUER_TOKENS: {
 			this.sendTokens(evtid);
 
 			break;
 		}
-		case MessageAction.GET_PROOF: {
+		case OutletAction.GET_PROOF: {
 			const token = this.getDataFromQuery("token");
 
 			requiredParams(token, "unsigned token is missing");
@@ -89,11 +86,13 @@ export class Outlet {
 
 			break;
 		}
-		case MessageAction.COOKIE_CHECK: {
+		case OutletAction.COOKIE_CHECK: {
 			this.sendMessageResponse({
 				evtid: evtid,
-				evt: MessageResponseAction.COOKIE_CHECK,
-				thirdPartyCookies: localStorage.getItem("cookie-support-check"),
+				evt: OutletResponseAction.COOKIE_CHECK,
+				data: {
+					thirdPartyCookies: localStorage.getItem("cookie-support-check"),
+				}
 			});
 
 			break;
@@ -167,9 +166,11 @@ export class Outlet {
 
 			this.sendMessageResponse({
 				evtid: evtid,
-				evt: MessageResponseAction.PROOF,
-				issuer: this.tokenConfig.tokenName,
-				proof: tokenProof,
+				evt: OutletResponseAction.PROOF,
+				data: {
+					issuer: this.tokenConfig.tokenName,
+					proof: tokenProof
+				}
 			});
 		} catch (e: any) {
 			logger(2,e);
@@ -183,21 +184,23 @@ export class Outlet {
 
 		this.sendMessageResponse({
 			evtid: evtid,
-			evt: MessageResponseAction.ISSUER_TOKENS,
-			issuer: this.tokenConfig.tokenName,
-			tokens: issuerTokens,
+			evt: OutletResponseAction.ISSUER_TOKENS,
+			data: {
+				issuer: this.tokenConfig.tokenName,
+				tokens: issuerTokens
+			}
 		});
 	}
 
 	public sendErrorResponse(evtid: any, error: string) {
 		this.sendMessageResponse({
 			evtid: evtid,
-			evt: MessageResponseAction.ERROR,
+			evt: ResponseActionBase.ERROR,
 			errors: [error],
 		});
 	}
 
-	public sendMessageResponse(response: MessageResponseInterface) {
+	public sendMessageResponse(response: ResponseInterfaceBase) {
 		// dont send Message if no referrer defined
 		if (!document.referrer) {
 			return;
